lisa.tests.scheduler.load_tracking: Use CPU from rt_app_profile for InvarianceItemBase._get_trace_signal()
`cpus = trace.analysis.tasks.cpus_of_tasks()` can't be used to get the
trace signal in the RqInvariance tests. `cpus` is created over the
un-cropped task trace, i.e. it might contain a CPU on which the task
was executing before rt-app could apply the taskset command on it.
This would lead to the fact that the CPU util signal comes from this
CPU as well which lets the test fail.

Use the `cpus` information of the `test` phase of the rt_app_profile
instead which will return the CPU the task should run on after the
taskset command is executed.

This fixes RqInvariance test failures seeing on juno-r2 and sd845
(1 out of 25 tests failed on one OPP).

Example on juno-r2:

Before:

    1@600000:
        actual max: 306
        actual mean: 268.0410731776392
        max error: 99.86727576622783 %      <-- !!!
        mean error: 4.721012409323003 %
        simulated max: 1023.9999999999974
        simulated mean: 314.66717783544584

After:

     1@600000:
        actual max: 306
        actual mean: 277.7188731660481
        max error: 0.8225020995934196 %    <-- !!!
        mean error: 0.1271045484900567 %
        simulated max: 303.1302706805885
        simulated mean: 276.4207643812954

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>